SWITCHYARD-2592 SCA cluster invocation doesn't work if both of consumer/producer are deployed onto same karaf node
Moved cache instance from activator to component so that all the SCA activator on a same runtime can share it
