+++ sca/src/main/java/org/switchyard/component/sca/deploy/SCAActivator.java
@@ -13,25 +13,14 @@
  */
 package org.switchyard.component.sca.deploy;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStream;
-
-import javax.naming.InitialContext;
 import javax.xml.namespace.QName;
 
 import org.infinispan.Cache;
-import org.infinispan.configuration.parsing.ConfigurationBuilderHolder;
-import org.infinispan.configuration.parsing.ParserRegistry;
-import org.infinispan.manager.DefaultCacheManager;
-import org.infinispan.manager.EmbeddedCacheManager;
-import org.jboss.logging.Logger;
 import org.switchyard.component.sca.RemoteEndpointPublisher;
 import org.switchyard.component.sca.SCAEndpoint;
 import org.switchyard.component.sca.SCAInvoker;
 import org.switchyard.component.sca.SCALogger;
 import org.switchyard.component.sca.SCAMessages;
-import org.switchyard.config.Configuration;
 import org.switchyard.config.model.composite.BindingModel;
 import org.switchyard.config.model.composite.SCABindingModel;
 import org.switchyard.deploy.BaseActivator;
@@ -45,54 +34,21 @@ import org.switchyard.remote.infinispan.InfinispanRegistry;
  */
 public class SCAActivator extends BaseActivator {
     
-    private static final String CACHE_CONTAINER_ROOT = "java:jboss/infinispan/container/";
-    private static final String CACHE_NAME_PROPERTY = "cache-name";
-    private static final String CACHE_CONFIG_PROPERTY = "cache-config";
-    private static final String DISABLE_REMOTE_TRANSACTION_PROPERTY = "disable-remote-transaction";
-    private static final String JGROUPS_CONFIG_PROPERTY = "jgroups-config";
     static final String[] TYPES = new String[] {"sca"};
 
-    private static Logger _log = Logger.getLogger(SCAActivator.class);
-    private Cache<String, String> _cache;
     private RemoteRegistry _registry;
     private RemoteEndpointPublisher _endpointPublisher;
     private boolean _disableRemoteTransaction = false;
 
     /**
      * Create a new RemoteActivator.
-     * @param environment component configuration used by the activator
+     * @param cache Infinispan cache
      */
-    public SCAActivator(Configuration environment) {
+    public SCAActivator(Cache<String, String> cache) {
         super(TYPES);
-        
-        // Determine the cache name
-        String cacheName = "cluster";
-        Configuration cacheNameConfig = environment.getFirstChild(CACHE_NAME_PROPERTY);
-        if (cacheNameConfig != null) {
-            cacheName = cacheNameConfig.getValue();
-        }
-        
-        // If a cache config is specified in the component configuration, then we 
-        // are managing the creation of the cache.  If not, then we look up a 
-        // cache manager and try to resolve the cache.
-        Configuration cacheFileConfig = environment.getFirstChild(CACHE_CONFIG_PROPERTY);
-        if (cacheFileConfig != null && cacheFileConfig.getValue() != null) {
-            Configuration jgroupsConfig = environment.getFirstChild(JGROUPS_CONFIG_PROPERTY);
-            createCache(cacheFileConfig.getValue(), cacheName, jgroupsConfig != null ? jgroupsConfig.getValue() : null);
-        } else {
-            lookupCache(cacheName);
-        }
-        
         // Configure a registry client to use the specified cache
-        if (_cache != null) {
-            _registry = new InfinispanRegistry(_cache);
-        } else {
-            SCALogger.ROOT_LOGGER.unableToResolveCacheContainer(cacheName);
-        }
-
-        Configuration bridgeRemoteTxConfig = environment.getFirstChild(DISABLE_REMOTE_TRANSACTION_PROPERTY);
-        if (bridgeRemoteTxConfig != null) {
-            _disableRemoteTransaction = Boolean.parseBoolean(bridgeRemoteTxConfig.getValue());
+        if (cache != null) {
+            _registry = new InfinispanRegistry(cache);
         }
     }
     
@@ -141,52 +97,11 @@ public class SCAActivator extends BaseActivator {
         return _endpointPublisher;
     }
     
-    private void createCache(String cacheConfig, String cacheName, String jgroupsConfig) {
-        ClassLoader origCl = Thread.currentThread().getContextClassLoader();
-        try {
-            InputStream configStream = null;
-            try {
-                File f = new File(cacheConfig);
-                if (f.exists() && f.isFile()) {
-                    configStream = new FileInputStream(f);
-                } else {
-                    configStream = SCAActivator.class.getClassLoader().getResourceAsStream(cacheConfig);
-                }
-
-                ClassLoader cacheClassLoader = DefaultCacheManager.class.getClassLoader();
-                Thread.currentThread().setContextClassLoader(cacheClassLoader);
-                ConfigurationBuilderHolder holder = new ParserRegistry(cacheClassLoader).parse(configStream);
-                if (jgroupsConfig != null) {
-                    holder.getGlobalConfigurationBuilder()
-                          .transport()
-                          .defaultTransport()
-                          .addProperty("configurationFile", jgroupsConfig);
-                }
-                _cache = new DefaultCacheManager(holder, true).getCache(cacheName);
-            } finally {
-                if (configStream != null) {
-                    try {
-                        configStream.close();
-                    } catch (Exception e) {
-                        e.fillInStackTrace();
-                    }
-                }
-           }
-        } catch (Exception ex) {
-            _log.debug("Failed to create cache for distributed registry", ex);
-        } finally {
-            Thread.currentThread().setContextClassLoader(origCl);
-        }
-    }
-    
-    private void lookupCache(String cacheName) {
-        // Attempt to resolve the cache container to use for the distributed registry through JNDI
-        try {
-            EmbeddedCacheManager cm = (EmbeddedCacheManager)
-                    new InitialContext().lookup(CACHE_CONTAINER_ROOT + cacheName);
-            _cache = cm.getCache();
-        } catch (Exception ex) {
-            _log.debug("Failed to lookup cache container for distributed registry", ex);
-        }
+    /**
+     * Set disalbe remote transaction.
+     * @param disable if true, disable remote transaction
+     */
+    public void setDisableRemoteTransaction(boolean disable) {
+        _disableRemoteTransaction = disable;
     }
 }
